Create input component and move handler functions to their proper components.
