+++ src/components/CurrencyInput/CurrencyInput.js
@@ -1,0 +1,17 @@
+const CurrencyInput = ({ input, setInput }) => {
+  const handleInputChange = event => {
+    const num = event.target.value
+
+    setInput(num.replace(/[^\d.]/g, ""))
+  }
+  return (
+    <input
+      placeholder="Enter the amount to convert..."
+      className="currency-input"
+      value={input}
+      onChange={handleInputChange}
+    />
+  )
+}
+
+export default CurrencyInput
